CRUD pembayaran

+++ application/controllers/Admin.php
@@ -8,7 +8,7 @@ class Admin extends CI_Controller {
 		$this->load->model('m_model');
 		$this->load->helper('my_helper');
         $this->load->library('upload');
-        if($this->session->userdata('logged_in')!=true) {
+        if($this->session->userdata('logged_in')!=true && $this->session->userdata('role') != 'admin') {
             redirect(base_url().'auth');
         }
 	}
@@ -28,46 +28,34 @@ class Admin extends CI_Controller {
         $this->load->view('admin/tambah_siswa', $data);
     }
 
-    public function aksi_tambah_siswa() {
-        $foto = $this->upload_img('foto');
-        if ($foto[0] === false) {
-            $data = [
-                'foto' => 'User.png',
-                'nama_siswa' => $this->input->post('nama'),
-                'nisn' => $this->input->post('nisn'),
-                'gender' => $this->input->post('gender'),
-                'id_kelas' => $this->input->post('id_kelas'),
-            ];
-            $this->m_model->tambah_data('siswa', $data);
-            redirect(base_url('admin/siswa'));
-        } else {
-            $data = [
-                'foto' => $foto[1],
-                'nama_siswa' => $this->input->post('nama'),
-                'nisn' => $this->input->post('nisn'),
-                'gender' => $this->input->post('gender'),
-                'id_kelas' => $this->input->post('id_kelas'),
-            ];
-            $this->m_model->tambah_data('siswa', $data);
-            redirect(base_url('admin/siswa'));
-        }
-    }
-
-    // Upload Image
-    public function upload_img($value)
+    public function aksi_tambah_siswa()
     {
+        $file_name = $_FILES['foto']['name'];
+        $file_temp = $_FILES['foto']['tmp_name'];
         $kode = round(microtime(true) * 1000);
-        $config[ 'upload_path'] = './images/siswa/'; 
-        $config[ 'allowed_types'] = 'jpg|png|jpeg';
-        $config['max_size'] = '30000';
-        $config['file_name'] = $kode;
-        $this->upload->initialize($config);
-        if ($this->upload->do_upload($value)) {
-            return array( false, '' );
+        $file_name = $kode . '_' . $file_name;
+        $upload_path = './images/siswa/' . $file_name;
+
+        if (move_uploaded_file($file_temp, $upload_path)) {
+        $data = [
+            'foto' => $file_name,
+            'nama_siswa' => $this->input->post('nama'),
+            'nisn' => $this->input->post('nisn'),
+            'gender' => $this->input->post('gender'),
+            'id_kelas' => $this->input->post('id_kelas'),
+        ];
+        $this->m_model->tambah_data('siswa', $data);
+        redirect(base_url('admin/siswa'));
         } else {
-            $fn = $this->upload->data();
-            $nama = $fn['file_name'];
-            return array( true, $nama );
+        $data = [
+            'foto' => 'User.png',
+            'nama_siswa' => $this->input->post('nama'),
+            'nisn' => $this->input->post('nisn'),
+            'gender' => $this->input->post('gender'),
+            'id_kelas' => $this->input->post('id_kelas'),
+        ];
+        $this->m_model->tambah_data('siswa', $data);
+        redirect(base_url('admin/siswa'));
         }
     }
 
@@ -97,9 +85,35 @@ class Admin extends CI_Controller {
         }
     }
 
-    public function hapus_siswa($id) {
-        $this->m_model->delete('siswa', 'id_siswa', $id);
-		redirect(base_url('admin/siswa'));
+    public function hapus_siswa($id) 
+    {
+        $siswa = $this->m_model->get_by_id('siswa', 'id_siswa', $id)->row();
+        if($siswa) {
+            if ($siswa->foto !== 'User.png'){
+                $file_path = './images/siswa/' . $siswa->foto;
+
+                if(file_exists($file_path)) {
+                    if(unlink($file_path)) {
+                        $this->m_model->delete('siswa', 'id_siswa', $id);
+                        redirect(base_url('admin/siswa'));
+                    } else {
+                        echo "Gagal menghapus file.";
+                    }
+                } else {
+                    echo "File tidak ditemukan.";
+                }
+            } else {
+                $this->m_model->delete('siswa', 'id_siswa', $id);
+		        redirect(base_url('admin/siswa'));
+            }
+        } else {
+            echo "Siswa tidak ditemukan.";
+        }
     }
+
+    // public function hapus_siswa($id) {
+    //     $this->m_model->delete('siswa', 'id_siswa', $id);
+	// 	   redirect(base_url('admin/siswa'));
+    // }
 }
 ?>
